drm/anx7808: Use interrupts for connection detection
This patch replaces the polling of the downstream connection with a
solution that is interrupt based. The driver will now turn on when there is
a cable plugged in. If there is no downstream connection, the driver
will wait for intp/dpcd interrupts to let it know when a downstream
monitor is available. At that point, the driver will assert hotplug.

BUG=chrome-os-partner:21538
TEST=Tested on spring with active & passive dongles

Change-Id: Ib4ad656795f74430e12778f003d1273b15673089
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168794