UPSTREAM: perf/x86: Fix uncore PCI fixed counter handling
There was a bug in the handling of SNB-EP/IVB-EP uncore PCI
fixed counters, e.g., IMC.

It would cause erratic values to be returned for the IMC
clockticks event. This was due to a bogus hwc->config value
which was then written to PCI config space.

The erratic values can be seen via:

  $ perf stat -a -C 0 -e uncore_imc_0/clockticks/ -I 1000 sleep 10

The fixed counter has most fields marked as reserved with
hw reset values of 0. Yet the kernel was defaulting to a
hwc->config = ~0 and that was causing the issues.

This patch sets the hwc->config values for fixed uncore event
to 0. Now, the values of IMC clockticks is correct.

BUG=chromium:474724
TEST=build/boot on Link

Change-Id: Id0dd2321ef3698bcc1c7672181edb0366e1c3556
Signed-off-by: Stephane Eranian <eranian@google.com>
Reviewed-by: Andi Kleen <ak@linux.intel.com>
Cc: peterz@infradead.org
Cc: zheng.z.yan@intel.com
Link: http://lkml.kernel.org/r/20130909195350.GA17643@google.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit dbc33f7016747bfce64f3d1e3af63f1251fcbf85)
Reviewed-on: https://chromium-review.googlesource.com/264477
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>