ARM: dts: Add trackpad info to kirby's device tree.
The language in the comments in this section was left intact even though it's
really intended for situations where we have more than one trackpad that we'll
probe for. I expect there's a decent change we'll have more than on trackpad
at some point, and this way we'll avoid rediscovering the problem mentioned in
the comments.

BUG=chrome-os-partner:22458
TEST=Built and booted on kirby.

Change-Id: Ie0cf7e57f203ebfbe9922579e6629d722dc3c60d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168689
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>