CHROMIUM: s5p-mfc: don't disable clock when next ctx is pending
In MFC's dynamic clock gating, we turn on the clock in try_run and
turn off the clock upon receiving an interrupt. But this
leads to excessive gating/ungating of the clocks in the case of
multi-instance video playback. The clock gets disabled in ctx1's irq
and then immediately turned on for ctx2's try_run.

A better solution is to turn off the clocks only when there are no new
frames to be processed in any context. This is done by conditionally
clocking on/off calls in try-run. clock-off is done outside try-run
only for suspend and error scenarios.

BUG=chrome-os-partner:22388
TEST=Boot on peach. Run multi-instance video for a long duration. No
spurious sysmmu faults are observed. Pause video and observe that mfc
is clock-gated.

Change-Id: Ib411999686f81ba01c8a46a7812b035461bdfe7d
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/167830
Tested-by: Arun Mankuzhi <arun@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>