mmc: dw_mmc: Check for card's presence during resume
It is possible that a card may be insterted when the system is in
suspend mode. Upon system resume, mmc subsystem is not aware of any such
insertion and hence it will not enumerate the such cards. Lets check for
card presence by scheduling card_work upon resume inorder to notify mmc
subsystem that a card was insterted/removed while asleep.

BUG=chrome-os-partner:13920
TEST=build and booted on pit.
TEST= booted system out of emmc, run rtcwake -m mem -s 10 to suspend the system
      and then insert SD card while asleep, observed that when system resume it has
      detected SD cards will all partitions
TEST= run suspend_stree_test multiple times

Change-Id: I0a1518e7d88ac7e0da4089d1e2915670bbd80d65
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/169095
Reviewed-by: Grant Grundler <grundler@chromium.org>