ASoC: samsung: add support for snow with max98090
This adds support for snow with the max98090 codec to the
daisy_max98095 machine driver.  The main difference is that
microphone bias is controlled by the pin "MICBIAS" on the
max98090 vs. "MICBIAS2" on the max98095.

BUG=chrome-os-partner:22590
TEST=Built and booted on snow-rev5 board.
Tested audio with browser as well as aplay utility.

Change-Id: Ifaec045c83f94995a7bb92c4e840bfe2cfb59eda
Signed-off-by: Padmavathi Venna <padma.v@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/169294
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>