usb: samsung-usb3phy: Ensure mutual exclusion during PHY tuning
The Samsung USB PHY driver needs to write a CR-port register as part of
its "PHY tuning" workaround for an Exynos5420 hardware bug. This is a
multi-stage register access which should not happen multiple times in
parallel, or wrong values could end up written to the wrong CR-port
address and cause weird errors later on. This patch uses the driver's
mutex to ensure that this access is protected.

BUG=chrome-os-partner:22674
TEST=Run suspend_stress_test and/or power_SuspendStress for a bazillion
times, make sure USB always comes up correctly.

Change-Id: Ied4548e85216c03d86445910fda69b8aba417fce
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169430
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>