cpufreq: exynos-iks: initialize secondary cpu policy based on cpu0
New cpufreq policies are created for the secondary CPUs when they
are brought up during resume, so if minimum or maximum frequencies
are adjusted prior to suspend (due to thermal throttling, for example),
they will be lost upon resume.  Since CPU0's policy is saved across
suspend/resume and throttling is usually applied to all CPUs, set
the min/max frequency of the secondary CPUs equal to that of CPU0
after resume.

BUG=chrome-os-partner:20290
TEST=Set scaling_max_freq to 700Mhz on all CPUs prior to suspend
and verify that scaling_max_freq is still set to 700Mhz for all
CPUs after resume.

Change-Id: I9229cee9eaf71b8c23e437cf2eee325d5a8aadb2
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169581
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>