ARM: dts: kirby: Leave the power to memory turned on during sleep.
Unlike on pit, the memory on kirby is powered off the max77802 PMIC. The
setting for the regulator it was on, buck5, were still set to turn off during
sleep, however, which would mean across a sleep/wake cycle we'd most likely
lose the contents of memory. Oddly we wouldn't always lose memory and would
still be able to still wake in some situations.

BUG=chrome-os-partner:22759
TEST=Built and booted on kirby. Suspended and resumed successfully.
BRANCH=None

Change-Id: Iac6b55b3e5208907ceed8b66fa02a4389633a096
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/169935
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>