CHROMIUM: usb: Warm reset USB3 ports that get stuck in Polling state
Some devices are not responding well to the USB3 port disable sequence
which will disable the link and then put it back into RxDetect state.
Instead the link gets stuck in the Polling state and does not recover.

This can happen after a suspend/resume cycle with certain USB3 devices
if the endpoint is renumbered and so a logical disconnect is attempted.
It can also happen at runtime with especially misbehaving devices.

The sequence of events for the more common suspend/resume case is:

[<ffffffffbad31a26>] hub_port_disable+0x74/0x23c
[<ffffffffbad31c0a>] hub_port_logical_disconnect+0x1c/0x32
[<ffffffffbad345c5>] usb_port_resume+0x2c5/0x38d
[<ffffffffbad2fcbc>] ? __each_dev+0x27/0x27
[<ffffffffbad42d4c>] generic_resume+0x1c/0x1e
[<ffffffffbad3b2cf>] usb_resume_both+0x94/0x106
[<ffffffffbad3bf5a>] usb_resume+0x1e/0x9b
[<ffffffffbad2fcbc>] ? __each_dev+0x27/0x27
[<ffffffffbad2fd03>] usb_dev_resume+0x47/0x4e

When this happens the USB3 Port Link State reads 0x2e0 instead of 0x2a0
and the inserted device is not detected.  In some cases removing and
re-inserting the device can cause it to come back, but other times it
will remain in polling state even when the device is removed.

This behavior happens regularly on the Haswell Chromebooks where we leave
ports enabled and powered in suspend in order to support wake from USB.
I had been attempting to work around the issue in firmware but that has
caused additional headaches like the link falling into compliance mode
and not recovering.

This change introduces an additional check into hub_usb3_port_disable()
function that will wait for the link to enter the RxDetect state and if
it fails to enter that state after a lengthy 1500ms timeout and is still
in the Polling state it will issue a warm port reset to recover the link.

BUG=chrome-os-partner:22818
TEST=manual: several suspend/resume cycles with various USB3 devices
attached to ensure that the devices are detected on resume.

Change-Id: I4af8f7f7d97459b1047694d538c163f3726d7652
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170171
Reviewed-by: Benson Leung <bleung@chromium.org>