ARM: dts: Add a "codec" node to the kirby device tree.
This node was present in the pit device tree and had been copied to the kirby
device tree, but it was removed in order to get the kernel booting. This
change adds it back in again since the sound configuration on kirby is the
same as on pit.

BUG=None
TEST=Booted with this device tree on kirby and, with other changes, played
sound on kirby from the headphone jack and from attached spring speakers.

Change-Id: I58a1428573757db7a5632dc1335c74d68dbf65e3
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170125
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>