UPSTREAM: Input: wacom - support EMR and MFT sensors of Cintiq Companion Hybrid
Adds support for the sensors integrated in to the Cintiq Companion
Hybrid. These sensors use by-and-large the same protocol as the
Cintiq 24HD touch.

NOTE: The ExpressKeys on the Cintiq Companion Hybrid are wired to
both the EMR controller and CPU GPIO pins. It may be necessary to
disable their functionality in this driver if building a custom
Android kernel for this device (lest two events be sent to userspace
for every button press).

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

TEST=manually
BUG=chromium:431322
Change-Id: Ib19b8fcd9571cbd496e19e1945a5d7574e2388ed
Reviewed-on: https://chromium-review.googlesource.com/250178
Tested-by: Jonathan Rockway <jrockway@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Jonathan Rockway <jrockway@chromium.org>