UPSTREAM: Input: wacom - send proper tablet state info when pen leaves proximity
The EMR sensor of Intuos4 and newer devices can send a packet with
0x20 in its second byte which contains only location information.
This packet is sent when the pen is too far away to reliably obtain
other information. Because only the location is updated userspace
will continue to see stale information for other values, most notably
TOUCH, PRESSURE, and DISTANCE.

This patch ensures userspace is not left with stale device state
information when the pen re-enters proximity (possibly leading to
the assumption the pen is still in contact even though it is
actually hovering).

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

TEST=manually
BUG=chromium:431322
Change-Id: I1aad69aa6eb089456601c78a802abe09d6e404fe
Reviewed-on: https://chromium-review.googlesource.com/250179
Tested-by: Jonathan Rockway <jrockway@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Jonathan Rockway <jrockway@chromium.org>