i2c: exynos5: consider bus busy as an error condition and reset
This patch makes the bus busy condition to be considered as an
error state (after 100ms interval) and resets the bus.

Also set the number of retries to 3 so that the i2c core retries
the transfer in that case.

BUG=chrome-os-partner:22834,chrome-os-partner:22837
TEST=Stress i2c bus on Pit.  For example:
  localhost ~ # while true; do i2cget -f -y 4 0x9 0x42; done
would occasionally produce i2c failures, but now it does not.

Change-Id: Id6b92998a6551cc8309ce2f42af3eea844a6a4f6
Signed-off-by: Naveen Krishna Chatradhi <ch.naveen@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170282
Reviewed-by: Doug Anderson <dianders@chromium.org>