cros_ec: stop calling ->cmd_xfer() directly
Instead of having users of the ChromeOS EC call the interface-specific
cmd_xfer() callback directly, introduce a central cros_ec_cmd_xfer()
to use instead.  This will allow us to put all the locking and retry
logic in one place instead of duplicating it across the different
drivers.

BUG=chrome-os-partner:22825
TEST=Communication with EC (flashrom, I2C tunnel, etc.) still works
on Pit.

Change-Id: Ib2aa4fd0df5854e8ec02d419ebac09b945a13525
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170746
Reviewed-by: Simon Glass <sjg@chromium.org>