ARM: dts: Disable the ADC on snow rev5
There are no thermistors on snow rev5.  Remove them from the common
file.  This means that we don't need to mark the ADC as disabled on
spring, so remove that.  BUCK5 is proper for VDD18_ADC0 on all
cros5250 boards, so leave it in the common file.

BUG=chrome-os-partner:22591
TEST=Thermistors work on rev4; they don't show up on rev5.
(test with grep "" /sys/class/hwmon/*/device/*_input)

Change-Id: I46b38c97000b0aa3a5e900f2955384908462abb4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previously-Reviewed-on: https://chromium-review.googlesource.com/170812
Reviewed-on: https://chromium-review.googlesource.com/170843
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>