mmc: dw_mmc: Add hardware lock error (HLE) to the CMD error list
Hardware locked error set when the dw_mmc controller cannot load a
command issued by software. When software sets the start_cmd bit in the
CMD register, the controller tries to load the command. If the command
buffer is already filled with a command, this error is raised. Currently
driver does not handle HLE interrupt, which results in a system hang.

HLE can be simulated by quickly inserting and removing sd card in card
slot (30 ~ 40 times in say 20 secs)

BUG=chrome-os-partner:22603
TEST=build and booted on pit. Simulated HLE error by quickly
inserting/removing sd card and put a print in interrupt handle to see if
HLE is set and with this patch see system not hanging any more, which
was used to hang with HLE is set and was not handled previously.

Change-Id: Id57eb2e087be7d548972bf5ce76ff029ce9b398c
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/170904
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>