pincntrl: add support for ams AS3722 pin control driver
The AS3722 is a compact system PMU suitable for mobile phones, tablets etc.

Add a driver to support accessing the GPIO, pinmux and pin configuration
of 8 GPIO pins found on the ams AS3722 through pin control driver and
gpiolib.

The driver will register itself as the pincontrol driver and gpio driver.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>

A few changes have happened in the kernel since 3.8. Rather than porting
all of them right now this patch backports this driver to support the
pinctrl framework as it is in 3.8.

Specifically, this handles the api change for pin_config_set from taking
all configs, to a single config. Also, this change changes the definitions
of pinctrl_ops, pinmux_ops, and pinconf_ops to non-const to match the
kernel 3.8 version of the pinctrl_desc structure.

BUG=None
TEST=Verfied that when enabled and given nodes, driver comes up

https://patchwork.kernel.org/patch/2975711/

Change-Id: Idf95df97633c146fb189a2d404417bcf4f0598cd
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/172011
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>