thermal: exynos: handle gate clock for misplaced TRIMINFO register
On Exynos5420, a TMU's misplaced TRIMINFO register may be located on
a TMU using a different gate clock.  Allow for a "tmu_apbif_triminfo"
clock to be specified in the device-tree which will be ungated when
accessing the TRIMINFO register.

BUG=chrome-os-partner:22774
TEST=Observe that the TRIMINFO_25 value is different for the GPU's
TMU before and after this change.  On my MP6 Pit it previously read
0 (resulting in abnormally high temperature readings), but now reads
56.

Change-Id: I11d252a55ad948e9090da77414a3c1208c45795f
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/171652
Reviewed-by: Doug Anderson <dianders@chromium.org>