BACKPORT: vfs: Add a function to lazily unmount all mounts from any dentry.
The new function detach_mounts comes in two pieces.  The first piece
is a static inline test of d_mounpoint that returns immediately
without taking any locks if d_mounpoint is not set.  In the common
case when mountpoints are absent this allows the vfs to continue
running with it's same cacheline foot print.

The second piece of detach_mounts __detach_mounts actually does the
work and it assumes that a mountpoint is present so it is slow and
takes namespace_sem for write, and then locks the mount hash (aka
mount_lock) after a struct mountpoint has been found.

With those two locks held each entry on the list of mounts on a
mountpoint is selected and lazily unmounted until all of the mount
have been lazily unmounted.

v7: Wrote a proper change description and removed the changelog
    documenting deleted wrong turns.

BUG=chromium:773474
TEST=build/boot on link, run xfsFilesystemTestSuiteExt4Plain

Change-Id: Ia92ae9dbfd6eb924e7b9990e673237a647787793
Signed-off-by: Eric W. Biederman <ebiederman@twitter.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
(cherry picked from commit 80b5dce8c59b0de1ed6e403b8298e02dcb4db64b)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
[SR: swich from lock_mount_hash() helpers back to old br_lock_write]
Reviewed-on: https://chromium-review.googlesource.com/762640
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>