ARM: dts: Remove the revision number from the snow model
It turns out that older versions of mosys are sometimes run on a newer
device.  Those older versions of mosys can't deal with the "Rev" at
the end of the string.  Remove the "Rev" from the rev4 dts file to
make old mosys happy.

Note that we won't need to worry about this in the future.  The rev5
boards can't run older images anyway and newer versions of mosys use
the "compatible" property instead of the "model" property.

BUG=chromium:301190
TEST=Boot and see that model is "Google Snow"

Change-Id: I8983600c57d0928b71b1362b8c86f5072c6ce7b0
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previously-Reviewed-on: https://chromium-review.googlesource.com/171696
Reviewed-on: https://chromium-review.googlesource.com/171765
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>