USB: dwc3: Add workaround for bulk transaction errors
DWC3 revisions from 1.90a to 2.10a have a bug.
For ss bulk-in data packet, when the host detects a
DPP error or the internal buffer becomes full,
it retries with an ACK TP Retry=1. However, under the following
conditions, the Retry=1 is falsely carried over to the next burst.
If there is only single active asynchronous SS EP at the time,
the active asynchronous EP is a Bulk IN EP and the burst with
the correctly Retry=1 ACK TP and the next burst belong to
the same transfer.

This patch is the workaround for this bug.

BUG=chrome-os-partner:23039
TEST=Tested Sandisk Extreme USB 3.0 16GB flash drive.
Video and audio play fine from it, and we don't see any sluggishness.

No more device reset warning messages as shown below appear:
[  209.157896] usb 2-1: reset SuperSpeed USB device number 2 using xhci-hcd
[  209.178114] usb 2-1: Parent hub missing LPM exit latency info.  Power management will be impacted.
[  209.186457] xhci-hcd xhci-hcd.4.auto: xHCI xhci_drop_endpoint called with disabled ep eeacfb00
[  209.194294] xhci-hcd xhci-hcd.4.auto: xHCI xhci_drop_endpoint called with disabled ep eeacfb2c

Also for confirming things see that the value at
GUCTL register of DWC3's global register space has a value:
localhost ~ # armio 0x1200c12c
1200c12c:7fc0c010
which is maintained across suspend/resume.

Change-Id: If97d2412da318d9156e1907df2d526d05ceb8d0f
Signed-off-by: Vivek Gautam <gautam.vivek@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/172237
Reviewed-by: Julius Werner <jwerner@chromium.org>