drm/exynos: Don't synchronously cancel hotplug work
This patch changes the hotplug work cancel from a synchronous one to
asynchronous. There is a race in the drm_helper_hpd_irq_event path where
if the hpd line drops low while we're in the helper functions,
the drm layer may try to power off hdmi from the worker's stack.

BUG=chrome-os-partner:22800
TEST=Couldn't reproduce the bug, but I'm confident this will fix it :)

Change-Id: I4dc3d2555399f7b389fba66e256ecdd8a74040d6
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172450
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>