mali: dvfs: Restore exynos5250 freq -> {gpll, aclk_400} lookup
For exynos5420, fout_gpll is programmed to the desired gpu frequency, and
sclk_g3d is set to the same by using a divider of 1.  This works because
the frequency levels in mali_dvfs_infotbl_exynos5420[] have been carefully
chosen to exactly match the VPLL frequencies in vpll_24mhz_tbl[]
(clk-exynos5420.c).

However, this is not a general solution, and is in fact broken for
exynos5250, where the frequencies in mali_dvfs_infotbl_exynos5250[] do
not match the GPLL frequencies in gpll_24mhz_tbl[] (clk-exynos5250.c).

This patch just restores the old behavior for exynos5250, which used a
switch-based lookup of gpll & aclk rates for the given target frequency.

A better, but much more intrusive fix may be to make use of the common
clock framework which has the ability to adjust a child clock's parent, if
configured properly.

Another potential fix would be to select GPU frequencies for exynos5250
that exactly correspond to valid GPLL outputs.  This could probably be
done easily, but requires coordination with samsung hardware engineering
to validate gpu frequencies, GPLL frequencies, and corresponding ASV and
DVFS voltages.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:303881
TEST=Per repro steps in chromium:303881, confirm sysfs reports the
  requested clock at each DVFS level.

Change-Id: I7a373cfd89aa1afe1f43a809ba36fe4bfea8427c
Reviewed-on: https://chromium-review.googlesource.com/172745
Reviewed-by: Akshu Agrawal <akshu@chromium.org>
Tested-by: Akshu Agrawal <akshu@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>