exynos: mcpm: Turn off APLL when A15 cluster is not active
When the A15 cluster is going down we can save power by powering off
the APLL and turning it back on when the A15 cluster is coming up.

BUG=chrome-os-partner:22219
TEST=Boot peach, run WebGL and see IKS working.
See that power is reduced by ~32mW on ARM rail and
~10mW on INT when checked along with the voltage reduction
patch.
Run memstress_tester and see there are no issues.

Change-Id: I904710f8d9a1deda3363c0c4f422e3a0f7520c4a
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173074
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>